+++ build.gradle
@@ -10,6 +10,9 @@ repositories {
 }
 
 dependencies {
+    implementation('org.springframework:spring-context:5.3.8')
+    testImplementation('org.springframework:spring-test:5.3.8')
+
     testImplementation 'org.junit.jupiter:junit-jupiter-api:5.7.0'
     testRuntimeOnly 'org.junit.jupiter:junit-jupiter-engine:5.7.0'
 }
