--- src/main/java/Main.java
@@ -1,11 +1,0 @@
-import logic.BubbleSort;
-
-import java.util.Arrays;
-
-public class Main {
-    public static void main(String[] args) {
-        BubbleSort<String> sort = new BubbleSort<>();
-
-        System.out.println("[result] " + sort.sort(Arrays.asList(args)));
-    }
-}
