+++ src/main/java/com/java/spring/config/Config.java
@@ -1,0 +1,9 @@
+package com.java.spring.config;
+
+import org.springframework.context.annotation.ComponentScan;
+import org.springframework.context.annotation.Configuration;
+
+@ComponentScan("com.java.spring")
+@Configuration
+public class Config {
+}
