+++ src/main/java/com/java/spring/logic/BubbleSort.java
@@ -1,10 +1,14 @@
-package logic;
+package com.java.spring.logic;
+
+import org.springframework.stereotype.Component;
 
 import java.util.ArrayList;
 import java.util.List;
 
-public class BubbleSort <T extends  Comparable<T>> {
+@Component
+public class BubbleSort <T extends  Comparable<T>> implements Sort<T> {
 
+    @Override
     public List<T> sort(List<T> list) {
         List<T> output = new ArrayList<>(list);
 
@@ -20,4 +24,5 @@ public class BubbleSort <T extends  Comparable<T>> {
 
         return output;
     }
+
 }
