+++ src/main/java/com/java/spring/logic/JavaSort.java
@@ -1,11 +1,15 @@
-package logic;
+package com.java.spring.logic;
+
+import org.springframework.stereotype.Component;
 
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
 
-public class JavaSort <T extends Comparable<T>>{
+@Component
+public class JavaSort <T extends Comparable<T>> implements Sort<T> {
 
+    @Override
     public List<T> sort(List<T> list) {
         List<T> output = new ArrayList<>(list);
         Collections.sort(output);
