+++ src/main/java/com/java/spring/logic/Sort.java
@@ -1,0 +1,7 @@
+package com.java.spring.logic;
+
+import java.util.List;
+
+public interface Sort <T extends Comparable<T>> {
+    public List<T> sort(List<T> list);
+}
