+++ src/main/java/com/java/spring/service/SortService.java
@@ -1,0 +1,22 @@
+package com.java.spring.service;
+
+import com.java.spring.logic.Sort;
+import org.springframework.beans.factory.annotation.Qualifier;
+import org.springframework.stereotype.Service;
+
+import java.util.List;
+
+@Service
+public class SortService {
+
+    private final Sort<String> sort;
+
+    public SortService(@Qualifier("bubbleSort") Sort<String> sort) {
+        this.sort = sort;
+        System.out.println("구현체 : " + sort.getClass().getName());
+    }
+
+    public List<String> doSort(List<String> list) {
+        return sort.sort(list);
+    }
+}
