+++ src/test/java/MainTest.java
@@ -1,3 +1,4 @@
+import com.java.spring.Main;
 import org.junit.jupiter.api.Test;
 
 public class MainTest {
