+++ src/test/java/com/java/spring/logic/JavaSortTest.java
@@ -1,4 +1,4 @@
-package logic;
+package com.java.spring.logic;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
@@ -7,16 +7,17 @@ import java.util.List;
 
 import static org.junit.jupiter.api.Assertions.*;
 
-class BubbleSortTest {
+class JavaSortTest {
 
-    @DisplayName("버블소트 - 리스트를 넣으면 정렬된 결과를 보여준다.")
+
+    @DisplayName("자바소트 - 리스트를 넣으면 정렬된 결과를 보여준다.")
     @Test
     void given_List_WhenExecuting_ThenReturnSortedList() {
         // Given
-        BubbleSort<Integer> bubbleSort = new BubbleSort<>();
+        JavaSort<Integer> javaSort = new JavaSort<>();
 
         // When
-        List<Integer> actual = bubbleSort.sort(List.of(3, 2, 4, 5, 1));
+        List<Integer> actual = javaSort.sort(List.of(3, 2, 4, 5, 1));
 
         // Then
         assertEquals(List.of(1, 2, 3, 4, 5), actual);
